Moved testing from routes/test, to tests/routes
- Moved files signuproutes.spec.js & authroutes.spec.js to tests/routes instead.
