+++ __tests__/routes/authRoutes.spec.js
@@ -1,10 +1,10 @@
 const request = require('supertest');
 const express = require('express');
-const router = require('../../routes/authRoutes'); // Import your router file here
+const router = require('../authRoutes'); // Import your router file here
 const connectDb = require('../../__tests__/fixtures/db')
 const makeFakeUser = require('../../__tests__/fixtures/fakeUser')
 const mongoose = require('mongoose');
-const { encrypt } = require('../../helpers/Password');
+const { encrypt } = require('../../helpers/password');
 
 const app = express();
 app.use(express.json());
@@ -28,6 +28,21 @@ describe('Login User route', () => {
     // Insert the fake user into the database
     await db.collection('users').insertOne(fakeUser);
   });
+  it('Returns error if user is not found', async () => {
+    const nonExistingUser = {
+      email: 'iDontExist@test.dk',
+      password: encrypt("12345678")
+    }
+
+    // Send a Post request to the login endpoint
+    const response = await request(`http://localhost:${PORT}`)
+      .post('/api/auth/login')
+      .send(nonExistingUser)
+      .expect(401);
+
+    // Verify the response body
+    expect(response.body.error.code).toBe("E0101");
+  });
 
   it('Returns error if user is not found', async () => {
     const nonExistingUser = {
@@ -39,10 +54,10 @@ describe('Login User route', () => {
     const response = await request(`http://localhost:${PORT}`)
       .post('/api/auth/login')
       .send(nonExistingUser)
-      .expect(404);
+      .expect(401);
 
     // Verify the response body
-    expect(response.body.message).toBe("User not found");
+    expect(response.body.error.code).toBe("E0101");
   });
 
 
@@ -58,9 +73,8 @@ describe('Login User route', () => {
       .expect(401);  
 
     // Verify the response body
-    expect(response.body.message).toBe("Incorrect password");
+    expect(response.body.error.code).toBe("E0105");
   });
-
 
   it('Returns token if user is found and password is correct', async () => {
     const correctCredentials = {
