+++ __tests__/routes/signupRoutes.spec.js
@@ -1,8 +1,8 @@
 const request = require('supertest');
 const express = require('express');
-const router = require('../../routes/signupRoutes'); // Import your router file here
-const connectDb = require('../fixtures/db')
-const makeFakeUser = require('../fixtures/fakeUser')
+const router = require('../signupRoutes'); // Import your router file here
+const connectDb = require('../../__tests__/fixtures/db')
+const makeFakeUser = require('../../__tests__/fixtures/fakeUser')
 const mongoose = require('mongoose');
 
 
@@ -33,7 +33,8 @@ describe('Signup User route', () => {
 
   it('Check that the endpoint saves the user in the database', async () => {
     const input = {
-      name: "test user",
+      firstName: "test user",
+      lastName: "test user",
       email: "test@email.com",
       password: "ABC123456!",
     };
@@ -52,7 +53,8 @@ describe('Signup User route', () => {
 
   it('Returns error if email is missing', async () => {
     const input = {
-      name: "test user",
+      firstName: "test user",
+      lastName: "test user",
       password: "ABC123456!",
     };
 
@@ -60,25 +62,31 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0208');
   });
 
 
   it('Returns error if password is missing', async () => {
     const input = {
-      name: "test user",
-      email: fakeUser.email,
+      firstName: "test user",
+      lastName: "test user",
+      email: fakeUser.email
     };
 
     const response = await request(`http://localhost:${PORT}`)
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0212');
   });
 
 
   it('Returns error if email does not contain @ and domain', async () => {
     const input = {
-      name: "test user",
+      firstName: "test user",
+      lastName: "test user",
       email: "testcase",
       password: "abc123456!",
     };
@@ -87,13 +95,16 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0206');
   });
 
 
   it('Returns error if email is less than 6 characters', async () => {
     const input = {
-      name: "test user",
-      email: "testcase",
+      firstName: "test user",
+      lastName: "test user",
+      email: "d@d.d",
       password: "abc123456!",
     };
 
@@ -101,6 +112,8 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0207');
   });
 
 
@@ -114,12 +127,14 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0209');
   });
 
-
-  it('Returns error if name contains illegal symbols', async () => {
+  it('Returns error if name is out of bounds (1-50)', async () => {
     const input = {
-      name: "test 😭😭😭",
+      firstName: "testtesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttesttest",
+      lastName: "testtesstteststtesttesttesttesttesttesttesttesttesttesttesttestteststtesttesttesttesttesttesttesttesttesttesttesttestteststtesttesttesttesttestte",
       email: "test@test.dk",
       password: "abc123456!",
     };
@@ -128,12 +143,32 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0210');
+  });
+
+
+  it('Returns error if name contains illegal symbols', async () => {
+    const input = {
+      firstName: "test 😭😭😭",
+      lastName: "test 😭😭😭",
+      email: "test@test.dk",
+      password: "abc123456!",
+    };
+
+    const response = await request(`http://localhost:${PORT}`)
+      .post('/api/signup/user')
+      .send(input)
+      .expect(400);
+
+      expect(response.body.error.code).toBe('E0211');
   });
 
 
   it('Test that the password is not stored as plain text', async () => {
     const input = {
-      name: "test user",
+      firstName: "hejsa",
+      lastName: "med digsa",
       email: "test@test.dk",
       password: "abc123456!",
     };
@@ -145,13 +180,14 @@ describe('Signup User route', () => {
 
     // Verify that the password is not stored as plain text
     const user = await db.collection('users').findOne({ email: input.email });
-    expect(user).not.toBe(input.password);
+    expect(user.password).not.toBe(input.password);
   });
 
 
   it('Test that emails must be unique when registering', async () => {
     const input = {
-      name: "test",
+      firstName: "test",
+      lastName: "test",
       email: fakeUser.email,
       password: "abc123456!",
     };
@@ -160,6 +196,8 @@ describe('Signup User route', () => {
       .post('/api/signup/user')
       .send(input)
       .expect(400);
+
+      expect(response.body.error.code).toBe('E0201');
 
   });
